Run initial filter of path expressions lazily.
Before this commit, running `<f><path>`
evaluated `<f>` strictly, then
evaluated indices of `<path>` lazily and ran the path for all outputs of `<f>`.
Now, this
evaluates the indices of `<path>` strictly, then
evaluates `<f>` lazily and runs the path for all outputs of `<f>`.

This enables, among others, writing the Fibonacci sequence as
`[0,1] | recurse([.[1], add])[0]`, which did not terminate before.

However, the new situation is not entirely satisfying,
because `<path>` might also be nonterminating.
However, the situation is similar to arithmetic operators,
where we also strictly evaluate the right-hand sides in order to
avoid repeating the same work for different left-hand side outputs.