Avoid a redundant allocation
`format!` allocates a string. If we pass that allocated string to `Error::str`, then the function will pass a reference of that string to `to_string` which will allocate it again and the first allocation will be freed.
This is because `to_string` only takes a reference.

Using `format_args!` instead of `format!` everywhere would be enough, but it would give the wrong impression that the function `Error::str` uses the ownership that is passed to it. Therefore, the signature is changed to only take a reference.