feat: teach alternation update-assignment operator
jq calls this the "alternative update-assignment operator".

note that jq documents the alternation operator as an arithmetic
operator and that jaq does not internally represent the alternation
operator as a math operator.

jq 1.6 considers chained `//=` operators without parentheses a syntax
error, like with `|=`; jaq allows chained `//=` operators by binding to
the right, at the same precedence as `|=`.