Revert replacement for RcIter.
This would have been an interesting idea to implement `inputs` without
storing the inputs in the context, by storing `RcIter` in the *filter*.
However, the problem here is that this implies that
when the input stream is dropped, the filter may not contain any more
reference to the input stream.
In practice, this is hard to achieve. Bloody hard. I gave up doing that.
Let's just keep inputs in the context --- it's ugly, but it works.