Evaluate RHS of `,` (concatenation) lazily.
I discovered the need for this via `cargo run -- -n '2 | recurse(. * 2)'`:
This panics at some point because of an integer overflow,
but most interestingly, its last output is
2305843009213693952, not
4611686018427387904, which is
the largest power of two that can be put into an i64.

That means that some operations, such as arithmetic,
are performed *before* some previous values have been output.
I then found out that even `., (. + 1 | .)` calculates
`. + 1` *before* outputting the input value (`.`).

I think that the route to take here is to
explicitly evaluate concatenation lazily.
On the other hand, this could make it possible to
evaluate a lot more things eagerly, where we use `once_with` today.
But let's not be hasty ...